reorganiza estrutura do backend e atualiza para type module
