+++ backend/src/routes/homeRoutes.js
@@ -1,0 +1,14 @@
+import { Router } from 'express';
+import {
+    listarDestaques,
+    listarNovidades,
+    listarAgendas
+} from '../controllers/homeController.js';
+
+const router = Router();
+
+router.get('/destaques', listarDestaques);
+router.get('/novidades', listarNovidades);
+router.get('/agendas', listarAgendas);
+
+export default router;
